backend/pyproject.toml
[tool.poetry]
name = "ultra-civic-backend"
version = "0.1.0"
description = "FastAPI backend for Ultra Civic carbon allowance retirement platform"
authors = ["Ultra Civic <paco@ultracivic.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlmodel = "^0.0.14"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.7"
alembic = "^1.12.1"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
slowapi = "^0.1.9"
resend = "^0.7.0"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
black = "^23.11.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pre-commit = "^3.6.0"
bandit = "^1.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
]